+++ src/components/common/QuestionTypeCountComponent.tsx
