+++ src/slices/examDataSlice.ts
@@ -1,17 +1,31 @@
-import {createSlice} from '@reduxjs/toolkit';
+import {createSlice, PayloadAction} from '@reduxjs/toolkit';
+import {Item} from "../type/Item";
 
-const initialState = {
+interface GroupedItem {
+    passageId: string | number;
+    passageUrl?: string | null;
+    items: Item[];
+}
+
+interface ExamDataState {
+    bookId: string | number | null;
+    totalQuestions: number | null;
+    groupedItems: GroupedItem[];
+    step1Data: any; // Replace `any` with a proper type if available
+}
+
+const initialState: ExamDataState = {
     bookId: null,
     totalQuestions: null,
     groupedItems: [],
-    step1Data: null
+    step1Data: null,
 };
 
 const examDataSlice = createSlice({
     name: 'examData',
     initialState,
     reducers: {
-        setExamData(state, action) {
+        setExamData(state, action: PayloadAction<ExamDataState>) {
             state.bookId = action.payload.bookId;
             state.totalQuestions = action.payload.totalQuestions;
             state.groupedItems = action.payload.groupedItems;
@@ -22,7 +36,7 @@ const examDataSlice = createSlice({
             state.totalQuestions = null;
             state.groupedItems = [];
             state.step1Data = null;
-        }
+        },
     },
 });
